requirements.txt
discord.py
discord.py[voice]
gTTS
youtube-dl
youtube_search
mutagen
python-dotenv
